Chaos refactoring n stuff
