+++ Choas/ChoasPlugin.cs
@@ -3,13 +3,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Choas.Features.OldMechanics;
 using Choas.SSSettings;
+using CustomCommands.Core;
+using LabApi.Events.CustomHandlers;
 using RedRightHand.CustomPlugin;
 using UserSettings.ServerSpecific;
 
 namespace Choas
 {
-    public class ChoasPlugin : CustomPluginCore
+    public class ChoasPlugin : CustomPluginCore<Config>
     {
         public override string Name => "Choas";
 
@@ -17,15 +20,34 @@ namespace Choas
 
         public override Version Version => new(1, 0, 0);
 
+        public override string ConfigFileName => "ChoasPlugin.yml";
+        
+        public override void LoadConfigs()
+        {
+            Config = LoadPluginConfigs<Config>();
+        }
+        
+		public CustomFeature[] features;
+        
+
         public override void Enable()
         {
-            CustomHandlersManager.RegisterEventsHandler(Events);
-        
-            if (ServerSpecificSettingsSync.DefinedSettings == null)
-                ServerSpecificSettingsSync.DefinedSettings = [];
+            if (Config.EnableCustomKeybinds)
+            {
+                CustomHandlersManager.RegisterEventsHandler(Events);
 
-            ServerSpecificSettingsSync.DefinedSettings = ServerSpecificSettingsSync.DefinedSettings.Concat(CustomSettingsManager.GetAllSettings()).ToArray();
-            ServerSpecificSettingsSync.SendToAll();
+                if (ServerSpecificSettingsSync.DefinedSettings == null)
+                    ServerSpecificSettingsSync.DefinedSettings = [];
+
+                ServerSpecificSettingsSync.DefinedSettings = ServerSpecificSettingsSync.DefinedSettings
+                    .Concat(CustomSettingsManager.GetAllSettings()).ToArray();
+                ServerSpecificSettingsSync.SendToAll();
+            }
+
+            features =
+            [
+                new OldMechanicsEvents(Config.EnableOldMechanics),
+            ];
         }
 
         public override void Disable() { }
