+++ Choas/Features/OldMechanics/OldMechanicsEvents.cs
@@ -1,18 +1,19 @@
-﻿using Choas.Components;
-using LabApi.Events;
+﻿using System.Linq;
+using Choas.Components;
+using CustomCommands.Core;
 using LabApi.Events.Arguments.PlayerEvents;
 using LabApi.Events.CustomHandlers;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UnityEngine;
+using LabApi.Features.Console;
 
-namespace Choas
+namespace Choas.Features.OldMechanics
 {
-    public class Events : CustomEventsHandler
+    public class OldMechanicsEvents : CustomFeature
     {
+        public OldMechanicsEvents(bool configSetting) : base(configSetting)
+        {
+            Logger.Info("What is this, 2019?");
+        }
+
         public override void OnPlayerSpawned(PlayerSpawnedEventArgs args) //Chance for two 939s to spawn and 173 speed tied to HP, like old times
         {
             if (args.Player.Team != PlayerRoles.Team.SCPs) return;
